requirements.txt
hydra-core==1.2.0
matplotlib==3.5.2
numpy==1.23.1
omegaconf==2.2.2
opacus==1.1.3
pandas==1.4.3
pytest==7.1.2
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
setuptools==61.2.0
tqdm==4.64.0
torch==1.11.0
torchvision==0.12.0
tensorboard==2.10.0